% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowsomlike_plot.R
\name{flowsomlike_plot}
\alias{flowsomlike_plot}
\title{Produce a FlowSOM-like plot}
\usage{
flowsomlike_plot(gng.fit, labels, max.size = 0.075)
}
\arguments{
\item{gng.fit}{The GNG produced by the \link{\code{gng}} function}

\item{labels}{Labels for the training samples}

\item{max.size}{The maximum size of visualised nodes}
}
\description{
Produce a FlowSOM-like plot
}
\examples{
# To be added soon!
}

