% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gng.R
\name{gng}
\alias{gng}
\title{Growing Neural Gas}
\usage{
gng(X, max.iter = 20000, epsilon.b = 0.05, epsilon.n = 0.001,
  age.max = 200, max.nodes = 30, lambda = 200, alpha = 0.5,
  beta = 0.99, make.projection = T, make.logs = F, verbose = T)
}
\arguments{
\item{X}{The input data}

\item{max.iter}{The max number of iterations}

\item{epsilon.b}{Move the winning node by epsilon.b times the distance}

\item{epsilon.n}{Move the neighbours of the winning node by epsilon.n times the distance}

\item{age.max}{Remove edges older than age.max}

\item{max.nodes}{The maximum number of nodes}

\item{lambda}{Insert new nodes every lambda iterations}

\item{alpha}{The decay parameter for error when a node is added}

\item{beta}{The decay parameter for error in every node every iteration}

\item{make.projection}{Project the points from X to a 2D plane}

\item{make.logs}{The decay for error of all nodes every iteration}

\item{verbose}{Will output progress if \code{TRUE}}
}
\value{
The GNG (need to be more specific in the future).
}
\description{
It is recommended to read the paper by Fritzke Bernd in order to gain a better insight into the given parameters.

Step 1, generating the input signal could be improved upon.
}
\examples{
# To be added soon!
}
\references{
Fritzke, Bernd. "A growing neural gas network learns topologies." Advances in neural information processing systems 7 (1995): 625-632.
}

