% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gng.R
\name{gng_project}
\alias{gng_project}
\title{Perform FR dimensionality reduction on GNG and project cells to same space}
\usage{
gng_project(gng_out, x = NULL)
}
\arguments{
\item{gng_out}{The object generated by \code{\link{gng}}}

\item{x}{If desired, the original dimred can also be projected using a randomForest}
}
\description{
Perform FR dimensionality reduction on GNG and project cells to same space
}
