% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gng.R
\name{plot_gng}
\alias{plot_gng}
\title{Produce a FlowSOM-like plot}
\usage{
plot_gng(
  gng_fit,
  plot_labels = NULL,
  plot_expression = gng_fit$node_space,
  max_size = 0.05,
  max_size_legend = 0.15
)
}
\arguments{
\item{gng_fit}{The GNG produced by the \code{\link{gng}} function}

\item{plot_labels}{Labels for the training samples. Is NULL, by default.}

\item{plot_expression}{Whether or not to plot the expression.}

\item{max_size}{The maximum size of visualised nodes}

\item{max_size_legend}{The maximum size of the legend nodes.}
}
\description{
Produce a FlowSOM-like plot
}
\examples{
data(iris)
gng_fit <- gng(x = as.matrix(iris[,1:4]))
plot_gng(gng_fit, plot_labels = iris[,5], max_size = 0.05)

}
