% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_gng.R
\name{project_gng}
\alias{project_gng}
\title{Perform FR dimensionality reduction on GNG and project cells to same space}
\usage{
project_gng(gng_out, x = NULL)
}
\arguments{
\item{gng_out}{The object generated by \code{\link[=gng]{gng()}}}

\item{x}{If desired, the original dimred can also be projected using a randomForest}
}
\description{
Perform FR dimensionality reduction on GNG and project cells to same space
}
