% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gng.R
\name{flowSOM_wrapper}
\alias{flowSOM_wrapper}
\title{FlowSOM wrapper}
\usage{
flowSOM_wrapper(data, colsToUse, nNodes = 100, nPasses = 50, ...)
}
\arguments{
\item{data}{fsom$FlowSOM$data as generated by \code{ReadInput}}

\item{colsToUse}{ids of the columns used}

\item{nNodes}{the number of nodes}

\item{nPasses}{the number of passes of the data}
}
\description{
FlowSOM wrapper
}

