% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowsomlike_plot.R
\name{flowsomlike_plot}
\alias{flowsomlike_plot}
\title{Produce a FlowSOM-like plot}
\usage{
flowsomlike_plot(gng_fit, labels, max.size = 0.075)
}
\arguments{
\item{gng_fit}{The GNG produced by the \code{\link{gng}} function}

\item{labels}{Labels for the training samples. Can also be NULL, if need be.}

\item{max.size}{The maximum size of visualised nodes}
}
\description{
Produce a FlowSOM-like plot
}
