% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gng.R
\name{gng_project}
\alias{gng_project}
\title{Perform FR dimensionality reduction on GNG and project cells to same space}
\usage{
gng_project(gng_out, make_projection = TRUE)
}
\arguments{
\item{gng_out}{The object generated by \code{\link{gng}}}

\item{make_projection}{Whether or not to also project the cells using a randomForest.}
}
\description{
Perform FR dimensionality reduction on GNG and project cells to same space
}
